+++ mcp_core/mcp_entry.py
@@ -1,6 +1,6 @@
 import sys
 import logging
-import core.config_core as config_core
+import server_core.config_core as config_core
 from mcp_core.server_setup import setup_mcp_server
 from mcp_core.hexstrike_client import HexStrikeClient
