+++ server_api/visual.py
@@ -1,7 +1,7 @@
 import time
 from flask import Blueprint, jsonify, request
 from datetime import datetime
-from core.process_manager import ProcessManager
+from server_core.process_manager import ProcessManager
 from visual.modern_visual_engine import ModernVisualEngine
 from intelligence.cve_intelligence_manager import CVEIntelligenceManager
