+++ server_core/__init__.py
@@ -3,10 +3,10 @@ HexStrike Core Module
 Shared infrastructure components for the HexStrike framework
 """
 
-from core.cache import HexStrikeCache
-from core.session_store import SessionStore
-from core.wordlist_store import WordlistStore
-from core.telemetry_collector import TelemetryCollector
+from server_core.cache import HexStrikeCache
+from server_core.session_store import SessionStore
+from server_core.wordlist_store import WordlistStore
+from server_core.telemetry_collector import TelemetryCollector
 from visual.modern_visual_engine import ModernVisualEngine
 
 __all__ = [
