+++ server_core/enhanced_command_executor.py
@@ -5,12 +5,12 @@ import subprocess
 import traceback
 from typing import Dict, Any
 from datetime import datetime
-import core.config_core as config_core
-from core.process_manager import ProcessManager
+import server_core.config_core as config_core
+from server_core.process_manager import ProcessManager
 from visual.modern_visual_engine import ModernVisualEngine
 
 # Global telemetry collector
-from core.telemetry_collector import TelemetryCollector
+from server_core.telemetry_collector import TelemetryCollector
 telemetry = TelemetryCollector()
 
 logger = logging.getLogger(__name__)
