+++ server_core/technology_detector.py
